new changes

+++ database/migrations/2022_06_22_142204_create_comments_table.php
@@ -13,17 +13,19 @@ class CreateCommentsTable extends Migration
      */
     public function up()
     {
-        Schema::create('comments', function (Blueprint $table) {
-            $table->id();
-            $table->string('name')->nullable();
-            $table->text('comment')->nullable();
-            $table->unsignedBigInteger('user_id')->nullable();
-            $table->string('status')->default(0);
-            $table->integer('question_id')->nullable();
-            $table->foreign('question_id')->references('id')->on('questions')->onDelete('cascade');
-            $table->softDeletes();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('comments')) {
+            Schema::create('comments', function (Blueprint $table) {
+                $table->id();
+                $table->string('name')->nullable();
+                $table->text('comment')->nullable();
+                $table->unsignedBigInteger('user_id')->nullable();
+                $table->string('status')->default(0);
+                $table->integer('question_id')->nullable();
+                $table->foreign('question_id')->references('id')->on('questions')->onDelete('cascade');
+                $table->softDeletes();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
